requirements.txt
# ═══════════════════════════════════════════════════════
#  JARVIS V2/V3 — Python Dependencies
#  Install with: pip install -r requirements.txt
# ═══════════════════════════════════════════════════════

google-genai>=1.0.0
requests>=2.31.0
